Разбивает файл на элементы
